feature(models): Add Commodity and Measure models to the Tariff tool.
Commodities and Measures are some of the core components for the Tariff tool. They
also form some of the most significant complexities. Commodities are designed in
life to follow a hierarchical system. Measures on the other hand have incredibly
complex relations pertaining to their conditions, duties and geographical areas.
This is combined with the need for a "time machine" allowing a user to see the state
of the database as it was at any point of time.

This commit creates the Commodity model as a hierarchical model using
Django-Treebeard. The commodities follow the Materialised Path design as it is
efficient for read operations (this dataset will be edited infrequently) and
allows multiple roots (each "section" for commodity codes would form a root). The
code also uses the default treebeard setup of a 4-character step length with a 36
character alphabet - giving each node the possibility of having roughly 1.6 million
children.

To facilitate the time machine system the design approach that has been decided on
is an append-only log style system - in that no row should be deleted, any
object that needs to be replaced should just be superseded with a new row. To
facilitate this a "valid_between" DateTimeRangeField has been added. This enables
the code to use database queries that find data as it was at specific dates with
ease. Some convenience methods implementing this have been added. An Exclusion
Constraint has also been added to enforce commodities with the same code cannot have
overlapping dates. This as a result requires the BTreeGIST extension installed on
postgresql - which is included in the migration. The system for enforcing
append-only updates has not yet been designed, however a concept of "successor"
and "predecessor" has been introduced, producing a link to the last used and next
used row for a commodity.

A simple Measure model has been added as well to showcase how this would connect
to a commodity as well and how it can be used in queries looking for measure in a
commodity tree.