Compile SCSS and Javascript assets in Dockerfile
The webapp now depends on NodeJS to fetch and compile GOV.UK Frontend and its dependencies.

This change introduces a multi-stage build, which first uses a NodeJS image to install GOV.UK Frontend, then passes the compiled assets to the Python image.

The container is now run with docker-compose, as it depends on a Postgres database.