Add footnote detail view [TP-88]
As a Tariff manager
I want to view the summary of a footnote
So that core information is visible at a glance

This change overrides the FootnoteViewSet to return a rendered Jinja template, which
uses GOV.UK Frontend Tabs and Table components to display the footnote core data.

The footnotes list is updated to make the footnote ID a link to the detail view.

This also adds a couple of test footnotes and footnote types in fixtures to be loaded
into the database with `python manage.py loaddata test-footnotes-fixtures`.