TP-123 Data model for regulations
Elements of the tariff are added, modified and removed by legal acts known as regulations. Regulations are published (by the EU in the Official Journal) and have a number and role type, which may be one of several, including Base regulation, Modification regulation and Abrogation regulation. TAMATO will only create and modify Base regulations, however we need to store historical data.

Regulations are stored in a single table (in contrast to the TARIC data model), with the relationships of the various types of regulation to each other represented as a number of reflexive many-to-many database relations.

A Modification regulation amends a single Base regulation.

A Prorogation regulation extends one or more Base or Modification regulations. Each regulation may be extended to different dates.

A Full Temporary Stop (FTS) regulation suspends one or more Base or Modification regulations. Each regulation is suspended to a single effective end date. A Partial Temporary Stop regulation is not recorded directly in the regulations table, but instead is referenced by the measures that are suspended.

An Abrogation regulation terminates a (Base/Prorogation/Modification/FTS) regulation. An Explicit Abrogation regulation specifies an effective date that the affected regulation is terminated. A Complete Abrogation regulation does not specify a date.

Also
* TP-133 Browse and find regulations
* TP-138 Core regulation data - Tab 1