Simplify WorkBaskets ready for TrackedModels
The data model for version tracking models removes the need for the WorkBasketItem model.

It has been decided that the approval process for WorkBaskets does not need to handle a submit/rejection/resubmit/approval cycle, removing the need for the ApprovalDecision model.

WorkBaskets are assigned a Transaction ID when they are submitted for approval. This is represented by the Transaction model, which has a One-to-One relation with WorkBasket.