TP-143: Create the data model and validations for geographical areas
All trade is carried out between geographical areas, however the
definition of a geographical area spans 3 concepts:

1) A Country
2) A Region (a trading area which is not recognised as a country)
3) A grouping of the above two concepts

Countries and Regions can be members of multiple groups whereas groups
follow somewhat of a hierarchical structure (there is a parent/child
system although this is rarely used).

This commit introduces the data model to cover geographical areas.
As all models linking to a geographical area are naive to what type
of area the area is (country, region or group) all these concepts are
stored in a single table. Meanwhile the group memberships are stored in
a through table so that memberships can be tracked separately of areas
with their own validity periods.