TP-117: Add system for tracking models and changes.
The project requires all changes to be tracked throughout the lifetime
of the system. It must be easy to query the latest versions of an object,
the currently active (based on a date) version of an object, as well as
to traverse it's entire history.

This commit introduces a TrackedModel base model. This is not an
abstract base model but instead is meant to be inherited concretely
making use of Django's Multi-Table Inheritance mechanism. Combined with
django-polymorphic this creates a polymorphic parent TrackedModel table
which efficiently fetches all the relevant children in any query
performed by the user.