TP-148: Add frontend view to browse geographical areas.
It is required for tariff managers to be able to view, browse and search
for geographical areas in order to edit and update them.

This commit introduces a frontend page for viewing and searching lists
of geographical areas. The search input utilises postgreSQLs full-text
search features and so can run a search over both the geographical areas
ID and its description set.