TP-166: Add frontend view to details of geographical areas.
It is required for tariff managers to be able to view, individual
geographical areas in order to edit ad update them.

This commit introduces a frontend page for viewing individual
geographical areas. This includes a tabbed view for core data,
descriptions, as well as a members/memberships tab.

The members/memberships tab changes based on the area_code. i.e.
if a geographical area is a group then it has members and so
displays the members tab. If a geographical area is not a group then it
does not have members but may have memberships - in which case these
are displayed in a memberships tab.