Make compatible with tamato-taric-import
The TAMATO TARIC importer script has a dependency on TAMATO's serializers and models to
validate and modify the TAMATO DB.

This change adds a `setup.py` to allow TAMATO to be added as a pip dependency

It also adds preliminary serializers for Commodity and Measure models, to allow the
importer to deserialize these models.

Finally, the ValiditySerializerMixin is modified to handle TARIC date format strings
(YYYY-MM-DD) as well as ISO 8601 format (the default), for convenience.