TP-189: adds the XML exporter API
The main consumers of this service require a TARIC3 compliant XML document
to be produced for any changes generated. These changes are generally grouped
at several levels. The top level is an `envelope`, which is the wrapper for the
document being sent. The next level is the `transaction`, which groups a block
of changes together - this is represented by the `WorkBasket` in this project.
The next group is the message which is just a wrapper around a record. The record
representing the individual object being changed.

This commit introduces an implementation to produce these compliant XML
documents. The idea behind this implementation is that, as all models
are children of `TrackedModel`, all models can follow the template method
design pattern. This enables us to have a standardised method for finding
and rendering templates relevant to each record. These record specific templates
can then in turn be included into a transaction template, building up the
TARIC document. In doing this the process offloads the rendering of each
individual model to the model itself. Giving a more granular and simplified way
to create and update how these records are presented. Thus in this case every
model is paired with a template based off of the model name (this is overrideable
with the use of either a `taric_template` attribute, or overriding the
`get_taric_template` method on the model. The template for the transaction
then adds the common transaction data, as defined in the workbasket and the
transaction model. Each record in the workbasket then has it's template included
with the data from the record inputted into it.

To facilitate this a PolymorphicModelSerializer is introduced. This requires
a mapping of models against their serializers. The mapping is managed by
using a decorator attached to the PolymorphicModelSerializer. When defining a
serializer for a model, decorating this serializer will automatically add the
serializer to the mapping, with it's relevant model.

Alongside this a new testing utility has been introduced, a decorator, which
will automatically generate an instance of the model required (via a factory)
and check if the TARIC XML generated from the model is valid against the
taric.xsd. This then returns the given XML for further testing if required.

Testing the generated data this way has shown that a couple tweaks were required
for several serializers and models, most notably the Regulation model,
these have been included in this commit.