TP-204: Add Commodity Code (Goods Nomenclature) data models
Commodities are the fundamental objects being imported into the country
that the tariff applies to. In the taric system these are generally
represented by a "Goods Nomenclature" within a hierarchical structure.
In this way commodities belong to categories and sub-categories. Where
various other constructs, such as measures, may apply to an individual
commodity, or to an entire category of commodities. So an easily
traversible tree is important.

This commit introduces a more robust data model to accurately represent
Goods Nomenclatures, including a description model, an association
to footnotes and an integrated tree.

In this process a Materialized Path hierarchical data structure is
used to represent the tree due to it's fast read times. This is
represented in the GoodsNomenclatureIndent class, with a foreign key
to the actual good. As a result the good can be edited separately of
its position within the tree.

Several legacy data structures have been wrapped up in the process.
Both the "origin" and "successor" tables weere used to represent
an auditable history of a commodity code, e.g. for when it was
split, or a new one added. As there are no dates related to this
it is a redundant table and can be simply represented by a single
self-referential "origin" foreign key. This way origins can be
found by following the origin foreign key, and successors can be
found by querying where an item is the origin.