TP-203: Edit footnotes and footnote descriptions
The main purpose of the TAMATO app is to provide a user interface for tariff managers to
edit the UK global tariff. So far, the user interface has been read-only.

This change adds the ability to edit the details of a Footnote, and the associated
description, and to save those changes in a current WorkBasket.

WorkBasket creation and submission is not in scope of this change, but the user is
prompted to choose a current workbasket if none is already open.

This change has highlighted the requirement to remove in-database exclusion constraints
on the models, due to draft changes necessarily conflicting with them. The business
rules that these constraints enforced are now implemented in validation code instead.