TP-257: Merge taric-importer into TaMaTo.
Currently the taric importer is separated from the main TaMaTo application but heavily relies on
the TaMaTo code. This commit combines the two projects so that the importing of specific
models can be handled close to where the models are defined, much in the same manner as the
exporter is handled.

The original repo was held at https://github.com/uktrade/tamato-taric-import/. The
majority of the code has been unchanged. However one important change has been made.
The individual importing requirements of each component is now decoupled from the rest of
the importer code. This is done by creating a registration mechanism on the ElementHandler.
This allows developers to simply decorate the component handler with a `register_component`
decorator from any parent component class, in doing so the current element will be included
as a potential child of the parent component.

E.g:

    class NormalChildComponent(ElementHandler):
        tag = Tag("normal-child")

    class ParentComponent(ElementHandler):
        tag = Tag("parent")
        normal_child = NormalChildComponent()

    @ParentComponent.register_component("registered_child")
    class RegisteredChildComponent(ElementHandler):
        tag = Tag("registered-child")

The above snipper will create a ParentComponent with the tag "parent", within this parent
will be both the "normal_child" component, but by using the `register_component` decorator
it will also include the "registered_child" component. Thus the code for the individual
 components is decoupled from the main importer code.

To further facilitate this the code now makes use of the `ready` hook provided on django
application configs. A common one has been provided which automatically tries to import any
import handlers in a package. The settings have been updated to ensure these configs are now
used. Without this the import_handlers files are never imported and the components therefore
don't register themselves.

The main command for importing has also been wrapped into a django management command. This
can be run with:

    python manage.py import_taric filename --status some_status --username some_username