Fix querysets with multiple versions of models
The CurrentWorkBasketMixin added multiple versions of the same model. For example when
editing footnotes: after editing a footnote for a second time, both the first and second
update versions were listed and attempting to browse to the detail view caused a 500
error as multiple objects were retrieved when only one was expected.

This change refactors the CurrentWorkBasketMixin, moving the queryset filtering code to
the TrackedModelQuerySet class, and ensuring that only the latest version of models in
the workbasket are added to the queryset.