Reuse existing draft in current workbasket
If there is already a draft of an object in the current workbasket, update it, rather
than creating a new draft.

This fixes the bug seen when editing the same footnote more than once. Multiple drafts
in the workbasket were added to the ModelView queryset, which caused an exception in the
DetailView.

In future, deleting an object should create or modify a draft in the workbasket with
UpdateType.DELETE