TP-206: Add importer for Goods Nomenclature and related models.
It is imperative that legacy data from the TARIC can be imported at will into the
TaMaTo database. Given the Goods Nomenclature models are now ready it is necessary
that the importer is setup for these models.

As this is the first set of models using the integrated importer there are a number
of changes made to the importing system to remove a few bugs and make it easier to
use the system for more models in the future.

A few discoveries of varying complexity have also been found. Firstly merged
models do not fit well with the current importer architecture. These instances have
been passed over to be done later as a way of adapting the importer to all models
affected by this (the majority of models in the system). The second is that, as
the hierarchical tree is largely unrepresented by the legacy TARIC system, this
will possibly require an entire implementation unto itself, again to be implemented
at a later date.

Tests have been added for all GN importer handlers - those which are currently not
working have been skipped.

Also updated the django requirements as polymorphic doesn't support 3.1 yet.