Add quotas data model
A major component of the UK tariff is quotas, or quantitative restrictions, which allow
goods to be imported at a modified tariff (usually lower) up to a certain volume or
value.  Quotas are used to balance protecting domestic producers of goods from
competition from imports with allowing exporters access to the market and allowing
domestic industries to import manufacturing resources.

An importer can specify which quota they wish to make use of by providing the Quota
Order Number. The order number is associated with one or more Quota Definitions, which
detail the volumes and periods for which the quota rate is available. The order number
is also associated with an origin (usually a country, or group of countries) to enable
setting different quota volumes and tariffs for different origins.

A quota can also be subdivided into sub-quotas, which may have varying validity periods,
origins or tariffs.

A quota can also be suspended, or blocked.

This PR introduces the data model for quotas and the business rules to support data
integrity.

In addition, a couple of common data model field types are introduced - NumericSID and
ShortDescription - which represent frequently used data types in the TARIC data model.

Also, the WorkflowStatus enum class is moved to prevent a circular dependency.
Unfortunately, this changed touched a lot of files.