Add quotas importer handlers
In order to read quota data from the TARIC3 XML seed and differential files we will
receive, we need to extend the TaMaTo importer command to parse those data from the XML,
validate them and write them to the database.

This PR adds parsers for TARIC3 formatted quota data and serializers for validation.