Reformat some long lines using the latest version of `black`
The latest `black` beta is more aggressive about
splitting across multiple lines.

We have chosen to stay updated to the latest
version for now so this commit applies those
reformats.

It also forces using the new version in our
requirements.txt so that there's no disparity
between what we run locally and in PR checks.