TP-299: Apply importer to Regulations models
This introduces a further extendable method in the handler for the
`save` step when actually saving to the database. The default
is to save through the serializer. This allows some finer changes to
be made when saving. This is useful in the instance of regulations
so that multiple models can be saved from a single handler.