Store abbreviations and prefixes for duty sentence objects
This commit adds human-readable abbreviations to
measurements and qualifiers and prefixes to duty
expressions. This will allow a duty sentence to be
both parsed and serialised using just the data in
the database alone – without requiring lookups.

E.g. humans tend not to write TNE but instead
"tonne" or "1,000 kg". We can store these in the
table and use them when interfacing with humans.

We may discover we need a separate table for
abbreviations if there are more than one per unit.