Set a default community code for regulations
The community code refers to what European
Community the TARIC measure is for. The EU doesn't
really keep track of this anymore and we certainly
are not all that interested in the UK tariff, but
the schema demands that it exists.

So let's set a sensible default for the community
code, and also include the options as a proper
enum for use elsewhere.