Split out envelope start and end for composition
This commit splits out the envelope start and end
XML snippets into separate templates to allow
anything that outputs XML to do so in a "streaming"
fashion – we no longer have to collect all of the
things we want to output in memory as we can just
output the entries one by one into a file.

We also add the option for specifying the envelope
ID as they need to match the filename. We need to add
something to the workbasket serializer to populate
this – we must revisit this in a later commit.

We also mess around with some spacing to hopefully
generate smaller XML files – after all, we have a
50MB limit!