Script for importing sheets of reference data
This is a script that can import and store a sheet
of reference data exported out of the old TAP. It
assumes that each row in one sheet directly
corresponds to one instance of a model in the new
schema. Anything more complex than that requires
a separate script.

The script isn't great really – it should probably
auto-detect the correct parsing functions from the
field definitions, but this is what we have for now.

Be aware of `validity_between` – it needs to be
passed to --tuples when importing a validity model.