Script for importing the UKGT
This is a script for importing the UKGT from a
spreadsheet. Things it does:

* Creates the draft regulation.
* End-dates any measures in an "old spreadsheet"
* Creates new measures from a "new spreadsheet"
* Parses duty expressions and creates components
* Detects trade remedy rates and keeps the old one
* Detects seasonal rates and creates measures that
  take account of the seasonality
* Generates one transaction per row
* Outputs everything into an XML and does NOT save
  any measures (yet!)

It assumes a certain schema for the spreadsheets
which has to be updated manually if they change.

It is "ready" to be converted to a more general
script as it should be possible to move `handle`
to a generic class and then specify different types
of imports by overriding `handle_row` and `compare_rows`.