Parser for duty sentences
This commit adds a parser for duty sentences. A
duty sentence is a combination of different types
of tariffs into one complete rate. Each duty
sentence is constructed of one or more duty
expressions that map directly to a single measure
component.

What is and is not a valid duty sentence is
controlled in part by the applicability codes
present on the duty expressions. This parser takes
account of these and won't parse e.g. a monetary
value being supplied for a Meursing component.

It also takes account of valid monetary units and
combinations of measurement units and qualifiers,
only allowing valid ones to be parsed.

It outputs unlinked measure components that can
be hooked up to a parent measure.