TP-240: Add Measures data models
Measures are the detail of the tariff. They specify how much must be paid to import a
specified good and how to calculate the amount, be it by percentage of value, by weight,
a fixed amount or varying by the origin location, date, or whether a quota applies.

This commit introduces the Django model classes and validators representing measures and
their associated business rules.

Many of the business rules are similar, and this commonality is factored out into
fixture functions in the conftest.py file.

Also in this commit are some refactoring of other business rules to make them more
consistent, and in the case of commodities, to make more effective use of Factory Boy
features.