TP-298: Implement importer for Geographical Areas.
It is a requirement for the project to be able to import Geographical Areas,
Descriptions and Memberships from legacy TARIC3 XML data.

This commit introduces the import parsers and handlers for these models.
One minor tweak was made to the models, introducing the area description
period SID as an SID on Descriptions. This is to aid in merging the description
period and description data together.