Remove dependency on freezegun
We are currently using a set of fixed dates for testing purposes.  Eventually the
current date will no longer fall within those date ranges.  This commit replaces the
fixed dates with relative dates, based on the current date at runtime.  This also
removes our dependency on freezegun for the few tests that needed it.