Handle regulation IDs from the national schema in historical data.
It appears that up until now the OJ numbers have
been allowed to be blank, and national IDs have
used different formats.

We will continue to support the different formats
(but should put more strict validation on our UI)
but the OJ numbers we will validate from this
point forwards.