Upgrde sheet importer to handle complex foreign keys and more types
We now allow multiple value foreign keys to be
used – as with the structured fields, columns on
the command line can be specified with bracketed
indexes. The index refers to one of the identifying
fields specified on the model, so that `name[0]`
will be the column containing values for the
`Name.identifying_fields[0]` column.

We also add more types – in particular,
ShortDescriptions are sometimes allowed to be
blank regardless of what the field says.