TP-333: Allow fetching of "latest" versions of a relation.
The data model within TaMaTo is intended to be versioned. That is to say
that any row within a table derived from TrackedModel holds some identifying
columns which are non-unique. If multiple rows have the same data within these
identifying columns they are seen as multiple versions of the same object.

However foreign keys must be made on the primary key field of a table. With
this being the case it is possible to create a row on one table and link it
to a row in another table. This row in another table can then be "updated" -
adding a new row to the table and a new version of the object. In this case
the foreign key may be considered stale, but only within the date range of
the new row that has been added.

To solve this problem a mechanism has been added to allow fetching of the
"latest" row representing the latest version of an object for all
relations pertaining to derivatives of TrackedModels.

In python this allows us to define a model like so:

```python3
class ExampleModel(TrackedModel):
    # must be a TrackedModel
    other_model = models.ForeignKey(OtherModel, on_delete=models.PROTECT)
```

And then access the latest version of the relation by simply adding `_latest`
to the attribute name like so:

```python3
example_model = ExampleModel.objects.first()
example_model.other_model_latest  # Gets the latest version
```

If done like in the above example, a query is generated specifically to fetch
this latest version of the row. However another mechanism has been added at
the queryset level which enable this to be done in a single query. This is
the `with_latest_links` QuerySet method.

This method finds all the relations of a model and for each one adds a column
populated by a SQL subquery that looks roughly like so:

```sql
SELECT Row_to_json({relation_table}) AS "json_data"
  FROM  {relation_table}
 INNER JOIN "common_trackedmodel" U1
    ON {relation_table}."trackedmodel_ptr_id" = U1."id"
 INNER JOIN "workbaskets_workbasket" U2
    ON U1."workbasket_id" = U2."id"
 WHERE {relation_table}.{identifying_field} = {outer_relation_join}.{identifying_field}
   AND U2."status" IN ({approved_statuses})
 ORDER BY {relation_table}."trackedmodel_ptr_id" DESC
 LIMIT 1
```

This caches the data for each row on the object instances in a dictionary
which is then used to create the related object in python when called upon.

To facilitate this a new `RowToJson` Func has been added, adding the ability
to call the postgres specific `row_to_json` function. This must be used in
conjunction with the `RowToJsonSubquery` which adds the proper arguments to
the function.