TP-309: Document biz rules with start dates
We have discovered that some of the business rules in TARIC have been
added to the system after data that breaks the rule has already been
present, and these have not been retroactively applied. This means that
in our implementation we need to give a free pass to data that was
inserted before the rule was formed.

The rules are:

    ME116: The quota order number on a measure must exist.
    Only applies to measures with a start date after 31/12/2007.

    ON6: The validity period of the geographical area must span the
    validity period of the quota order number origin.
    Only applies to quota order numbers with a start date after
    2006-12-31.

    ON9: When a quota order number is used in a measure then the validity
    period of the quota order number must span the validity period of the
    measure.
    Only applicable for measures with start date after 31/12/2007.

    ROIMB8: Explicit dates of related measures must be within the validity
    period of the base regulation.
    Only applicable to measure with start date after 31/12/2003.

This commit is to make sure these dates are reflected in code and
documentation.