Test helper for biz rules with cutoff dates
Several business rules are only applicable after specified cutoff dates,
for example ME116 is only applicable to measures with a start date after
31/12/2007.

This commit adds a test helper decorator function which runs a business
rule test once expected it to pass, then uses freezegun to set the
current time to before the specified start date and runs the test again,
this time expecting it to fail.

This saves duplicate code and verbose setting of validity periods on
dependent models in the business rule tests. It also brings the cutoff
date to the top of the test code for the business rule, making it more
noticeable and the tests easier to read.

To enable this decorator, the `Dates` class has been refactored to
replace the class constants with instance properties which generate date
ranges at run-time, rather than at import-time. This allows freezegun to
affect these date ranges. Code which uses the `date_ranges` fixture
needs no changes, but code which used the `Date` class directly has been
updated.