TP-342: Refactor GoodsNomenclatureIndents.
Further investigation has found that, by virtue of the loose
relations held in the legacy TARIC system, any Goods Nomenclature
Indent can hold multiple parent indents over its lifetime without
requiring an update. This is due to how parent relations are
worked out - all that is needed is for a new indent to be made
connected to the right item ID at the right depth for this to happen.

As a result of this, to try and stay moderately aligned to the TARIC
system and not require multiple updates to maintain the tree, it
has been deemed necessary to further abstract away the tree system.

This commit does this by refactoring the GoodsNomenclatureIndent model
to not be part of a hierarchical tree structure. Instead the tree is
moved to a non-tracked model with a foreign key to the indent. This
model is called GoodsNomenclatureIndentNode. The advantage of this is
to have a flexible tree which can be changed to reflect the correct
hierarchy for queries, but without the need to impact any outputs
to clients.