Importer for autonomous suspensions sheet
We now need to get arbitrary context from old rows,
not just the measure type, so the slicer has been
updated to offer up all old rows that match the CC.
This is so we can find footnotes that were present
on the old rows.

We have also factored out more utils and we now
force that we do not import any "inherited" rows –
input rows should always be whole measures and not
ancestor measures, becuase the latter are too
tricky to deal with and we don't really need them.